Add script for generating RPQBench queries
This patch is a follow up of the 62783f95 ('Add basic RPQBench dataset
information'). It adds script for generating RPQBench queries.

The provided queries don't contain source/destination vertices. Let's
add a script that allows generating random sources/destinations based on
the query templates and the original graph.

It can be used as follows.
```bash
./gen-queries ./sparql-queries-templates.txt rpqbench-trunc.nt
```